\name{build_folder_structure}
\alias{build_folder_structure}
\title{Build the folder structure for legacy data}
\usage{
build_folder_structure(project_name)
}
\arguments{
  \item{project_name}{the name of the project that you wish to create.}
}
\description{
Builds the data legacy folder structure in the users working directory.
}
\examples{
build_folder_structure(project_name="Test")
}
